+++ test/bootstrap.php
@@ -13,7 +13,6 @@ $files = [
 foreach ($files as $file) {
     if (file_exists($file)) {
         $loader = require $file;
-        $loader->addPsr4('Rx\\Websocket\\Test\\', __DIR__);
         break;
     }
 }